% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoutjoy.R
\name{plot.scoutjoy}
\alias{plot.scoutjoy}
\title{Plot SCOUTJOY results}
\usage{
\method{plot}{scoutjoy}(fit, label = "outlier", colors = c("darkred",
  "darkblue", "darkorange", "black"), se_bars = TRUE, legend = TRUE,
  legend_names = c("all variants", "excluding outliers", "outliers"),
  se_length = 0.025, ...)
}
\arguments{
\item{fit}{output of \link{scoutjoy}`}

\item{label}{string, whether to label "outliers", "all", or "none".}

\item{colors}{list of colors to use for plotting base regression, regression excluding outliers, outlier points, and other points, respectively.}

\item{se_bars}{logical, whether to include standard error bars}

\item{legend}{logical, whether to include a legend}

\item{legend_names}{list of names to use for slopes and outliers (only affects legend)}

\item{se_length}{length for caps on SE intervals; passed to \link{arrows}}

\item{...}{additional parameters to be passed directly to \link{plot()}}
}
\description{
Plot SCOUTJOY results
}
