% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoutjoy.R
\name{scoutjoy}
\alias{scoutjoy}
\title{Significant Cross-trait OUtliers and Trends in JOint York regression (SCOUTJOY)}
\usage{
scoutjoy(BetaOutcome, BetaExposure, SdOutcome, SdExposure, data,
  ids = NULL, CovIntercept = 0, ExposureIntercept = 1,
  OutcomeIntercept = 1, OutlierNull = "estimate",
  SignifThreshold = 0.05, NullReplicates = NA, maxOutlierReps = 10,
  seed = NULL, printProgress = TRUE)
}
\arguments{
\item{BetaOutcome}{String, giving the column name in `data` for the effect sizes (betas) on the outcome (y-axis trait)}

\item{BetaExposure}{String, giving the column name in `data` for the effect sizes (betas) on the exposure (x-axis trait)}

\item{SdOutcome, }{String, giving the column name in `data` for the standard error (se) of the effect sizes on the outcome (y-axis trait)}

\item{SdExposure, }{String, giving the column name in `data` for the standard error (se) of the effect sizes on the exposure (x-axis trait)}

\item{data}{Data frame with the specified column names containing effect sizes and standard errors}

\item{ids}{(Optional) string giving the column name in `data` of variant IDs to use for labeling results tables}

\item{CovIntercept}{Default = 0,}

\item{ExposureIntercept}{Default =1,}

\item{OutcomeIntercept}{Default =1,}

\item{OutlierNull}{Default = "estimate"}

\item{SignifThreshold}{Default = 0.05,}

\item{NullReplicates}{Default = NA,}

\item{maxOutlierReps}{Default =10,}

\item{seed}{Default = NULL,}

\item{printProgress}{Default = TRUE}
}
\value{
List with elements:
\describe{

  \item{\code{Global Test}}{xxx}
  
  \item{\code{xxx}}{xxx}

}
}
\description{
xxx
}
\details{
xxx
}
\examples{

x <- stats::rnorm(20)

}
\references{
xxx
}
